third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/scroll-snap-align-expected.txt

This is a testharness.js-based test.
[FAIL] Can set 'scroll-snap-align' to the 'none' keyword: none
  assert_class_string: expected "[object CSSStyleValue]" but got "[object CSSKeywordValue]"
[FAIL] Can set 'scroll-snap-align' to the 'start' keyword: start
  assert_class_string: expected "[object CSSStyleValue]" but got "[object CSSKeywordValue]"
[FAIL] Can set 'scroll-snap-align' to the 'end' keyword: end
  assert_class_string: expected "[object CSSStyleValue]" but got "[object CSSKeywordValue]"
[FAIL] Can set 'scroll-snap-align' to the 'center' keyword: center
  assert_class_string: expected "[object CSSStyleValue]" but got "[object CSSKeywordValue]"
Harness: the test ran to completion.
